% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{income_emilia}
\alias{income_emilia}
\title{The income of municipalities of "Emilia Romagna".}
\format{A data frame with 334 rows and 7 variables:
\describe{
  \item{municipality_code}{identification municipality code}
  \item{municipality}{name of the municipality}
  \item{province}{province of the municipality}
  \item{numtaxpay}{number of taxpayers in the municipality}
  \item{tot_inc}{average income of the municipality}
  \item{x_coord}{coordinate x of the municipality}
  \item{y_coord}{coordinate y of the municipality}
         }}
\source{
The dataset is a rearrangement from the data released by MEF - Dipartimento delle Finanze (Italy).
}
\usage{
income_emilia
}
\description{
The dataset contains the total income of the municipalities in the region "Emilia Romagna", in Italy. Each municipality is defined by their own ISTAT (Istituto nazionale di statistica, Italy) code, a name and its geographical positions (through a pair of coordinates).
For each of them we have the following auxiliary variables: province, number of taxpayers and total income of the municipality.
}
\keyword{datasets}
