% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simul2}
\alias{simul2}
\title{Simulated Population 2.}
\format{A data frame with  1000 rows and 11 variables:
\describe{
  \item{x}{coordinate x}
  \item{y}{coordinate y}
  \item{z21}{first value of the unit}
  \item{z22}{second value of the unit}
  \item{z23}{third value of the unit}
  \item{z24}{fourth value of the unit}
  \item{z25}{fifth value of the unit}
  \item{z26}{sixth value of the unit}
  \item{z27}{seventh value of the unit}
  \item{z28}{eighth value of the unit}
  \item{z29}{ninth value of the unit}
         }}
\source{
\insertRef{BIMJ:BIMJ1785}{Spbsampling}
}
\usage{
simul2
}
\description{
The dataset contains a simulated georeferenced population of dimension \eqn{N=1000}.
The coordinates are generated in the range \eqn{[0,1]} as a simulated realization of a particular random point pattern: the Neyman-Scott process with Cauchy cluster kernel.
The nine values for each unit are generated according to the outcome of a Gaussian stochastic process, with an intensity dependence parameter \eqn{\rho=0.01} (that means medium dependence) and with a spatial trend \eqn{x_{1}+x_{2}+\epsilon}.
}
\keyword{datasets}
