% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{income_emilia}
\alias{income_emilia}
\title{The income of municipalities of "Emilia Romagna".}
\format{A data frame with 334 rows and 7 variables:
\describe{
  \item{municipality_code}{identification municipality code}
  \item{municipality}{name of the municipality}
  \item{province}{province of the municipality}
  \item{numtaxpay}{number of taxpayers in the municipality}
  \item{tot_inc}{average income of the municipality}
  \item{x_coord}{coordinate x of the municipality}
  \item{y_coord}{coordinate y of the municipality}
         }}
\source{
The dataset is a rearrangement from the data released by
the Italian Finance Department, MEF - Dipartimento delle Finanze (Italy).
}
\usage{
income_emilia
}
\description{
The dataset contains the total income of the municipalities in the region
"Emilia Romagna", in Italy, for the year 2015. Each municipality is defined
by their own ISTAT (Istituto nazionale di statistica, Italy) code and a name.
For each municipality there are the following auxiliary variables: province,
number of taxpayers and spatial coordinates (geographical position).
}
\keyword{datasets}
