% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spbsampling.R
\docType{package}
\name{Spbsampling}
\alias{Spbsampling}
\alias{Spbsampling-package}
\title{Spatially balanced sampling designs.}
\description{
Selection of spatially balanced samples.
In particular, the implemented sampling designs allow to select probability samples spread over the population of interest, in any dimension and using any distance function (e.g. Euclidean distance, Manhattan distance).
The implementation has been done in \code{C++} through the use of \code{Rcpp} and \code{RcppArmadillo}.
}
\references{
\insertRef{BIMJ:BIMJ1785}{Spbsampling}

\insertRef{fast_selection}{Spbsampling}
}
\author{
Francesco Pantalone, Roberto Benedetti, Federica Piersimoni

Maintainer: Francesco Pantalone \email{pantalone.fra@gmail.com}
}
